refactor: extract Linear watcher settings fields to satisfy max-lines
